LayoutTests/media/modern-media-controls/airplay-button/airplay-button-expected.txt

Testing the AirplayButton class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS airplayButton.element.localName is "button"
PASS airplayButton.element.classList.contains("icon") is true
PASS airplayButton.element.classList.contains("airplay") is true
PASS airplayButton.iconName is "airplay"
PASS airplayButton.element.style.webkitMaskImage.includes("macOS/airplay@") is true
PASS macOSFullscreenAirplayButton.element.style.webkitMaskImage.includes("macOS/airplay-fullscreen@") is true
PASS iOSAirplayButton.element.style.webkitMaskImage.includes("iOS/airplay@") is true
PASS successfullyParsed is true

TEST COMPLETE
